demo: refactor example blocks
